+++ web/src/router/index.ts
@@ -11,6 +11,7 @@ const routes = [
   { path: '/api-endpoints', component: () => import('../views/ExternalApiView.vue') },
   { path: '/chat', component: () => import('../views/ChatView.vue') },
   { path: '/knowledge', component: () => import('../views/KnowledgeView.vue') },
+  { path: '/plugins', component: () => import('../views/PluginsView.vue') },
   { path: '/cleanup', component: () => import('../views/CleanupView.vue') },
   { path: '/help', component: () => import('../views/HelpView.vue') },
 ]